perf(core): optimize template expansion with early exit and allocation reuse
- Add early return in `expand()` when content has no template markers
- Skip `replace()` calls when key not present in string
- Rewrite `expand_env_vars` to use zero-copy scanning instead of repeated allocations
- Pre-allocate result buffer with `String::with_capacity`

+++ crates/y-core/src/agent.rs
@@ -47,7 +47,7 @@ pub struct DelegationOutput {
     /// The text output produced by the delegated agent.
     pub text: String,
     /// Approximate tokens consumed during the delegation.
-    pub tokens_used: u32,
+    pub tokens_used: u64,
     /// Input tokens consumed (for diagnostics breakdown).
     pub input_tokens: u64,
     /// Output tokens generated (for diagnostics breakdown).
@@ -198,7 +198,7 @@ pub struct AgentRunOutput {
     /// Text output produced by the agent.
     pub text: String,
     /// Tokens consumed during execution (input + output combined).
-    pub tokens_used: u32,
+    pub tokens_used: u64,
     /// Input tokens consumed (for diagnostics breakdown).
     pub input_tokens: u64,
     /// Output tokens generated (for diagnostics breakdown).
